requirements.txt
certifi
chardet
cycler
decorator
Django
idna
kiwisolver
lxml
matplotlib
networkx
nltk
numpy
pandas
pandas-datareader
Pillow
pyparsing
python-dateutil
pytz
requests
scipy
seaborn
six
sqlparse
urllib3
virtualenv
wrapt
